changed to arch based systems

+++ install.sh
@@ -1,38 +1,41 @@
 #!/bin/bash
-cd /home/tim/ || exit
+DIR="$(pwd)"
+cd "$HOME" || exit
 
-echo "--- installing via apt ---"
-packages=("zsh" "build-essential" "cmake" "ninja-build" "gettext" "unzip" "zsh" "nodejs")
+echo "--- updating system ---"
+sudo rm -rf /etc/pacman.conf
+sudo ln -s "$DIR/etc/pacman.conf" /etc/
+sudo pacman -Syyu --noconfirm --needed
+
+packages=(
+	"zsh"
+	"base-devel"
+	"zsh"
+	"nodejs"
+	"neovim"
+	"github-cli"
+	"go"
+	"google-chrome"
+	"npm"
+)
 
 for i in "${packages[@]}"; do
-	if ! [ -x "$(command -v $i)" ]; then
-		echo "--- installing $i ---"
-		sudo apt -y install $i
+	if ! [ -x "$(command -v "${i:?}")" ]; then
+		echo "--- installing ${i:?} ---"
+		yay -S "${i:?}" --noconfirm --needed
 	else
-		echo "--- $i already installed --- "
+		echo "--- ${i:?} already installed --- "
 	fi
 done
 
-echo "--- installing rust ---"
-curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y
-
-echo "--- installing golang---"
-wget https://go.dev/dl/go1.21.1.linux-amd64.tar.gz
-sudo rm -rf /usr/local/go && sudo tar -C /usr/local -xzf go1.21.1.linux-amd64.tar.gz
-rm -rf go1.21.1.linux-amd64.tar.gz
-export PATH=$PATH:/usr/local/go/bin
-
-if ! [ -x "$(command -v nvim)" ]; then
-	echo "--- installing neovim ---"
-	git clone https://github.com/neovim/neovim
-	cd neovim && make CMAKE_BUILD_TYPE=RelWithDebInfo
-	cd build && cpack -G DEB && sudo dpkg -i nvim-linux64.deb
-	rm -rf /home/tim/neovim
+if ! [ -x "$(command -v cargo)" ]; then
+	echo "--- installing rust ---"
+	curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y
 else
-	echo "--- neovim already installed --- "
+	echo "--- rust already installed --- "
 fi
 
-if ! [ -d "/home/tim/.oh-my-zsh" ]; then
+if ! [ -d "$HOME/.oh-my-zsh" ]; then
 	echo "--- installing oh-my-zsh ---"
 
 	chsh -s /usr/bin/zsh
@@ -46,36 +49,33 @@ cargo install tree-sitter-cli ripgrep bottom --locked
 go install github.com/jesseduffield/lazygit@latest
 go install github.com/dundee/gdu/v5/cmd/gdu@latest
 
-echo "--- installing github-cli ---"
-type -p curl >/dev/null || (sudo apt update && sudo apt install curl -y)
-curl -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo dd of=/usr/share/keyrings/githubcli-archive-keyring.gpg &&
-	sudo chmod go+r /usr/share/keyrings/githubcli-archive-keyring.gpg &&
-	echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/githubcli-archive-keyring.gpg] https://cli.github.com/packages stable main" | sudo tee /etc/apt/sources.list.d/github-cli.list >/dev/null &&
-	sudo apt update &&
-	sudo apt install gh -y
-
 echo "--- creating config ---"
+etc=(
+	"pacman.conf"
+)
+for i in "${etc[@]}"; do
+	sudo rm -rf /etc/"${i:?}"
+	sudo ln -s "$DIR/etc/${i:?}" /etc/
+done
+
 home=(
 	".gitconfig"
 	".zshrc"
 )
-
 for i in "${home[@]}"; do
-	rm -rf /home/tim/"${i}"
-	ln -s /home/tim/.dotfiles/home/"${i}" /home/tim/
+	rm -rf "$HOME/${i:?}"
+	ln -s "$DIR/home/${i:?}" "$HOME"
 done
-
-if ! [ -d "/home/tim/.config" ]; then
-	mkdir /home/tim/.config
-fi
 
 config=(
 	"nvim"
 )
-
+if ! [ -d "/home/tim/.config" ]; then
+	mkdir /home/tim/.config
+fi
 for i in "${config[@]}"; do
-	rm -rf /home/tim/.config/"${i}"
-	ln -s /home/tim/.dotfiles/config/"${i}" /home/tim/.config/
+	rm -rf /home/tim/.config/"${i:?}"
+	ln -s "${DIR}/config/${i:?}" /home/tim/.config/
 done
 echo "--- config created ---"
 echo "--- SETUP DONE ---"
